services/hawkeye/requirements.txt
ultralytics
redis
psycopg2-binary
opencv-python-headless
numpy
